Deprecate `convert_multiplier` replaced by simpler `promote_multiplier`
Taking into account the element type of the destination in
`convert_multiplier` turns out to be a bad idea.  A more simple method
`promote_multiplier` is provided as a replacement.  This open the path
to easier extension of methods to complex valued multipliers.

Fix `EndomorphismType` which is in fact `MorphismType`.