Rename `I` as `Id` and always export `Id`
The exported constant `I = Identity()` has been renamed as `Id` to avoid
conflicts with standard `LinearAlgebra` package.  `Id` is systematically
exported while `I` was only exported if not already defined in the `Base`
module.  The constant `LinearAlgebra.I` and, more generally, any instance
of `LinearAlgebra.UniformScaling` is recognized by LazyAlgebra in the
sense that they behave as the identity when combined with any LazyAlgebra
mapping.