Provide `DecoratedMapping`, make `unveil` more strict, use accessors
`DecoratedMapping` is the union of decorated mappings, that is `Adjoint`,
`Inverse` and `InverseAdjoint`.

Method `unveil` has a better defined behavior: for a decorated mapping, it
yields the embedded mapping; for other LazyAlgebra mappings (including
scaled ones), it returns its argument; for an instance of
LinearAlgebra.UniformScaling, it returns the equivalent LazyAlgebra
mapping (that is λ⋅Id).  To get the mapping embedded in a scaled mapping,
call the `unscaled` method.  These distinctive behaviors help to slarify
and simplify the code.

Attempt to rely on accessors in the code and avoid expressions like
`obj.key` except in accessors of course.