Simplify forbidding some constructor calls
A call to a constructor should always yield an instance of the
corresponding type.  Some calls to constructors are forbidden because
they should yield an instance of a different type if simplification
rules were applied.  This was implemented with traits, this now done
more simply by outer constructors.

According to this, calling Adjoint, Inverse or InverseAdjoint on the identity is
forbidden.