Simplify `vcombine` and `vcombine!`
Use neutral numbers avoiding dispatching again on the values of the
multipliers in `unsafe_vcombine!`. Use stage object to get rid of
`dispatch_vcombine!` and simplify dispatching by dispatching on `α` in a
1st stage and then on `β` in a 2nd stage.

Benchmarking shows that the new implementation is as fast, or even
faster, than the old one despite it does not calls `vcopy!`, `vzeros!`,
or `vscale!` but always `unsafe_vcombine!` in the end.