Relax number of dimensions
In top-level methods which check axes of their arguments, the signature
put no constraints on the number of dimensions (a `DimensionMismatch`
exception will be thrown anyway). In the lower-level methods (stage ≥ 1)
which assume that axes have been checked, we impose that the number of
dimensions be the same in the signature as a minimal check.