Convert and dispatch at different stages
Multipliers are factorized out form product of operators and the
conversion of multipliers is done before dispatching on their values.
This is to operate on non-scaled and non-product operators as early as
possible and to dispatch on multipliers that are numerically equal to 0
or 1 at the given precision (not before conversion).