Fix sums of 2 terms
Due to distribution of scalar multiplication among the terms of a sum
when simplifying sums, sums of 2 terms like `2*(A + B) - (B + A)` must
undergo the same processing than sums of more than 2 terms. The
specialization of `simplify` for sums of 2 terms has thus been removed.
To not penalize too much the computation time, sorting 2 terms has been
optimized.