Same `unsafe_vmul!` for sparse operators and their adjoint
For sparse operators, the `getindex` method takes care of conjugating
the returned value so it is not needed to have special versions of
`unsafe_vmul!` for the adjoint of sparse operators. All that's matter is
whether the operator must be accessed in a CSR, CSC, or COO order. There
are no regressions.