Replace (input|output)_axes(A) by (input|output)_shape(A)
This is to let the implementation decide how is specified a known i/o
shape (dimensions and or index ranges). This also clarifies what needs
to be implemented when `(Input|Output)HasShape` indicates a known shape.

Confusing `(col|row)_(size|axes|shape)` have been removed. That is
`col_size(A)` may be understood as the size of the columns of `A` while
it is the size of its rows! Such ambiguities are avoided with
`(input|routput)_(size|axes|shape)`.